ai-server/pyproject.toml
[project]
name = "jenny-ai"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Todd Cullen", email = "todd@100-x.ai"},
]
dependencies = [
    "fastapi>=0.115.5",
    "llm-mindmeld>=0.1.1",
    "uvicorn>=0.32.0",
    "pytest-asyncio>=0.24.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pytest.ini_options]
pythonpath = [
    "src"
]
